requirements.txt
ccxt>=4.0.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
PyYAML>=6.0
seaborn>=0.12.0
scipy>=1.10.0
plotly>=5.15.0
tables>=3.8.0
pyarrow>=10.0.0
requests>=2.28.0
python-dotenv>=1.0.0
